refactor into components, services, utils, and pages
