+++ components/clustering/__init__.py
@@ -1,0 +1,3 @@
+"""
+UI components for the clustering.
+"""
