+++ services/__init__.py
@@ -1,0 +1,3 @@
+"""
+Server-side business logic for the embedding explorer.
+"""
