--- ui/__init__.py
@@ -1,3 +1,0 @@
-"""
-UI components for clustering page.
-"""
