--- ui/clustering/__init__.py
@@ -1,3 +1,0 @@
-"""
-UI components for the clustering page.
-"""
