--- ui/shared/__init__.py
@@ -1,3 +1,0 @@
-"""
-Shared UI components for the embedding explorer.
-"""
